refactor: rewrite register API
affects: @logicflow/core
