+++ packages/core/src/view/edge/index.ts
@@ -1,0 +1,7 @@
+import BaseEdge from './BaseEdge';
+import LineEdge from './LineEdge';
+import PolylineEdge from './PolylineEdge';
+import BezierEdge from './BezierEdge';
+import Arrow from './Arrow';
+
+export { BaseEdge, LineEdge, PolylineEdge, BezierEdge, Arrow };
