+++ packages/core/src/view/node/index.ts
@@ -1,0 +1,17 @@
+import BaseNode from './BaseNode';
+import RectNode from './RectNode';
+import CircleNode from './CircleNode';
+import PolygonNode from './PolygonNode';
+import DiamondNode from './DiamondNode';
+import EllipseNode from './EllipseNode';
+import TextNode from './TextNode';
+
+export {
+  BaseNode,
+  RectNode,
+  CircleNode,
+  PolygonNode,
+  DiamondNode,
+  EllipseNode,
+  TextNode,
+};
